refactor!: remove radiotext event type
