+++ src/ingest/auth/refresh.ts
@@ -1,14 +1,14 @@
 import type { Context } from 'hono'
 import type { JwtPayload } from 'jsonwebtoken'
-import type { AuthRefreshBody } from '../../../schemas/auth.ts'
+import type { AuthRefreshBody } from '../../schemas/auth.ts'
 import { getNow } from '@frytg/dates'
 import logger from '@frytg/logger'
-import { firebaseRefreshToken } from '../../../utils/firebase/refresh-token.ts'
-import { getSafeHeaders } from '../../../utils/get-safe-headers.ts'
-import { badRequest as responseBadRequest } from '../../../utils/response/bad-request.ts'
-import { responseInternalServerError } from '../../../utils/response/internal-server-error.ts'
-import { responseOk } from '../../../utils/response/ok.ts'
-import { getValidatedBody } from '../../../utils/validation/zod-validate.ts'
+import { firebaseRefreshToken } from '../../utils/firebase/refresh-token.ts'
+import { getSafeHeaders } from '../../utils/get-safe-headers.ts'
+import { badRequest as responseBadRequest } from '../../utils/response/bad-request.ts'
+import { responseInternalServerError } from '../../utils/response/internal-server-error.ts'
+import { responseOk } from '../../utils/response/ok.ts'
+import { getValidatedBody } from '../../utils/validation/zod-validate.ts'
 
 const source = 'ingest/auth/refresh'
