+++ src/ingest/auth/reset.ts
@@ -1,12 +1,12 @@
 import type { Context } from 'hono'
-import type { AuthResetBody } from '../../../schemas/auth.ts'
+import type { AuthResetBody } from '../../schemas/auth.ts'
 import logger from '@frytg/logger'
-import { firebaseSendPasswordResetEmail } from '../../../utils/firebase/send-password-reset-email.ts'
-import { getSafeHeaders } from '../../../utils/get-safe-headers.ts'
-import { badRequest as responseBadRequest } from '../../../utils/response/bad-request.ts'
-import { responseInternalServerError } from '../../../utils/response/internal-server-error.ts'
-import { responseOk } from '../../../utils/response/ok.ts'
-import { getValidatedBody } from '../../../utils/validation/zod-validate.ts'
+import { firebaseSendPasswordResetEmail } from '../../utils/firebase/send-password-reset-email.ts'
+import { getSafeHeaders } from '../../utils/get-safe-headers.ts'
+import { badRequest as responseBadRequest } from '../../utils/response/bad-request.ts'
+import { responseInternalServerError } from '../../utils/response/internal-server-error.ts'
+import { responseOk } from '../../utils/response/ok.ts'
+import { getValidatedBody } from '../../utils/validation/zod-validate.ts'
 
 const source = 'ingest/auth/reset'
